Initialize Jupyter Book 2 with MyST
Starting with a minimal Jupyter Book initialized with `jupyter book init --write-toc`. Changed the `myst.yml` to use a proper title, description and the DevSeed logo. Included a <br> fix for the CC-BY-4.0 license in the main README.md file. Also added a JupyterBook badge to the main README.md.